+++ changelog.md
@@ -88,6 +88,7 @@
 - Shotgun Pellets are now Parriable
 - Added Large Animated Titles
 - Added Text Box Titles
+- Punching a Player using a Shield with a Knuckleblaster will now disable their Shield and give the `+ GUARD BREAK` style bonus
 ## Settings & Config
 - Removed almost all gamerules (only remaining is `ultra-startWithPiercer`)
 - Added Option to disable Screenshake